ALL DEVICE V.29 20221108 (成) 1. 將透傳工作時間計入休眠週期中以簡化週期循環 2. 改善休眠時間超過10分鐘後可能會沒有作用並造成睡眠週期異常的問題。 3. 修改Skynet在市電模式下的讀取週期判斷     - 所有Server都是市電供電且全都是Skynet模式，每2秒讀取一次Skynet內容     - 所有Server都是市電供電但其中含有逶傳模式時，以所設定的通訊週期(預設為60秒)讀取Skyent內容 4. 改善Client在逶傳模式中當Modbus指令間隔小於透傳暫存觸發時間時，可能會造成回應錯位的問題。 5. 透傳暫存觸發時間預設值由500ms改為300ms

+++ Mesh_Node.h
@@ -32,7 +32,7 @@
 typedef struct _NodeStageInfo
 {
     uchar   Stage;
-    uint16  Timer;
+    uint32  Timer;
 }NodeStageInfo,*PNodeStageInfo;
 
 typedef struct _NodeHeader_
@@ -536,7 +536,7 @@ extern uchar    NodeRole;
 extern PNodeStageInfo pStageInfo;
 extern PNodeEventInfo  pNodeEventInfo;
 extern uchar CountErr;
-extern uint16 GetInfoCycle;
+extern uint32 GetInfoCycle;
 extern uchar   UsartRxCount;   // Receive data from Rx bytes
 #include "Mesh_Client.h"
 #include "Mesh_Server.h"
